Starting work on Ratings. Rearranging IQDB into own package and separate from Boorus.
